elk/enrichers/requirements.txt
kafka-python==2.0.2
geoip2==4.7.0
cachetools==5.3.1
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3
prometheus-client==0.19.0